Add QINFO as what QSTATUS used to be, make QSTATUS be machine readable
This commit adds QINFO, which is essentially what QSTATUS used to be but now
even prettier, which its columns lining up nicely between lines. QSTATUS is now
a nicely machine readable version of that same data returned by QINFO, and in
fact QINFO uses QSTATUS as the source of its data. In doing these two changes
some large changes to the commands_test.go file were done, in effect making it
use QSTATUS instead of QINFO in all the places where it needed to check the
overall status of one or more queues.